Add files via upload
Problem Statement
Students are often worried about their chances of admission in graduate school. The aim of this machine learning model is to help students in shortlisting universities with their profiles. The predicted output gives them a fair idea about their admission chances in a particular university. This will assist students to know in advance if they have a chance to get accepted.

Data Definition
GRE Score: Graduate Record Exam (GRE) score. The score will be out of 340 points (numeric)

TOEFL Score: Test of English as a Foreigner Language2 (TOEFL) score, which will be out of 120 points (numeric)

University Rating: University Rating (Uni.Rating) that indicates the Bachelor University ranking among the other universities. The score will be out of 5 (numeric)

SOP: Statement of purpose (SOP) which is a document written to show the candidate's life, ambitious and the motivations for the chosen degree/ university. The score will be out of 5 points (numeric)

LOR: Letter of Recommendation Strength (LOR) which verifies the candidate professional experience, builds credibility, boosts confidence and ensures your competency. The score is out of 5 points (numeric)

CGPA: Undergraduate GPA (CGPA) out of 10 (numeric)

Research: Research Experience that can support the application, such as publishing research papers in conferences, working as research assistant with university professor (either yes or no) (categorical)

Chance of Admit: One dependent variable can be predicted which is chance of admission, that is according to the input given will be ranging from 0 to 1 (numeric).